+++ src/expr/dot.rs
@@ -29,7 +29,8 @@ pub struct DotVisitor<'a>(pub &'a DataType);
 
 impl<'a> Visitor<'a, DataType> for DotVisitor<'a> {
     fn column(&self, column: &'a Column) -> DataType {
-        self.0[column.clone()].clone()
+        //self.0[column.clone()].clone()
+        todo!()
     }
 
     fn value(&self, value: &'a Value) -> DataType {