+++ CHANGELOG.md
@@ -8,9 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.2.2] - 2023-08-29
+### Changed
 - Updated sqlparser version
 - Deactivate graphviz display by default
 - Deactivate multiplicity testing by default
+### Added
+- Implemented `DataType::hierarchy` method
 
 ## [0.2.1] - 2023-07-26
 ### Added